@@ -11,7 +11,7 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.4",
         "ext-json": "*",
         "becklyn/assets-bundle": "^2.3",
         "symfony/dependency-injection": "^4.1 || ^5.0",
@@ -25,7 +25,7 @@
         "twig/twig": "^2.10 || ^3.0"
     },
     "require-dev": {
-        "becklyn/rad-bundle": "^7.0.3",
+        "becklyn/rad": "^8.1.3",
         "roave/security-advisories": "dev-master",
         "symfony/browser-kit": "^5.0",
         "symfony/css-selector": "^5.0",
@@ -34,7 +34,8 @@
         "symfony/var-dumper": "^5.0"
     },
     "suggest": {
-        "becklyn/rad-bundle": "To use the 'form' dummy content."
+        "becklyn/rad": "To use the 'form' dummy content.",
+        "becklyn/rad-bundle": "To use the 'form' dummy content for projects that haven't upgraded to `becklyn/rad` yet."
     },
     "config": {
         "sort-packages": true
