@@ -4,7 +4,7 @@ namespace Tests\Becklyn\GluggiBundle\App;
 
 use Becklyn\AssetsBundle\BecklynAssetsBundle;
 use Becklyn\GluggiBundle\GluggiBundle;
-use Becklyn\RadBundle\BecklynRadBundle;
+use Becklyn\Rad\BecklynRadBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
 use Symfony\Bundle\TwigBundle\TwigBundle;
